Move setLockDuration to settings store
